# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Business Validation Platform Document API
  version: 0.12.0-phase3-taskrun
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >
    Canonical authentication, document upload, asynchronous processing,
    structured-plan mutation, and explicit confirmation contract.
servers:
  - url: /api/v1
...
security:
  - platformAuth: []
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List the current user's active projects
      operationId: listProjects
      x-implementation-phase: Phase 5
      responses:
        '200':
          description: Projects ordered by most recently updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectListApiResponse'}
        '401': {$ref: '#/components/responses/StandardError'}
    post:
      tags: [Projects]
      summary: Create a business validation project
      operationId: createProject
      x-implementation-phase: Phase 5
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateProjectRequest'}
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectDetailApiResponse'}
        '400': {$ref: '#/components/responses/StandardError'}
        '401': {$ref: '#/components/responses/StandardError'}
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: Get a project owned by the current user
      operationId: getProject
      x-implementation-phase: Phase 5
      responses:
        '200':
          description: Owned project detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectDetailApiResponse'}
        '401': {$ref: '#/components/responses/StandardError'}
        '404':
          description: Project does not exist or is not visible to the current user
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
...
components:
  securitySchemes:
    platformAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Phase 3 access JWT. The development X-User-Id adapter is profile-bound
        and is deliberately excluded from the public contract.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {type: integer, format: int64}
...
  responses:
...
    StandardError:
      description: Standard API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
  schemas:
...
    Meta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: {type: string}
        timestamp: {type: string, format: date-time}
    ApiError:
      type: object
      required: [code, message, fieldErrors, retryable]
      properties:
        code: {type: string}
        message: {type: string}
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}
        retryable: {type: boolean}
    ErrorApiResponse:
      type: object
      required: [success, error, meta]
      properties:
        success: {type: boolean, const: false}
        error: {$ref: '#/components/schemas/ApiError'}
        meta: {$ref: '#/components/schemas/Meta'}
...
    CreateProjectRequest:
      type: object
      required: [title]
      properties:
        title: {type: string, minLength: 1, maxLength: 150}
        description:
          type: [string, 'null']
          maxLength: 10000
        industryCategory:
          type: [string, 'null']
          maxLength: 100
    ProjectStatus:
      type: string
      enum: [DRAFT, ACTIVE, PAUSED, COMPLETED, ARCHIVED]
    ProjectStage:
      type: string
      enum:
        - DOCUMENT
        - STRUCTURING
        - LEGAL_REVIEW
        - FEASIBILITY
        - FINANCIAL
        - PERSONA_CONFIGURATION
        - PANEL_SURVEY
        - PANEL_DISCUSSION
        - REPORT
        - MARKETING
        - COMPLETED
    ProjectSummary:
      type: object
      required: [id, title, stage, status, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64}
        title: {type: string}
        industryCategory: {type: [string, 'null']}
        stage: {$ref: '#/components/schemas/ProjectStage'}
        status: {$ref: '#/components/schemas/ProjectStatus'}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/ProjectSummary'
        - type: object
          required: [ownerId, version]
          properties:
            ownerId: {type: integer, format: int64}
            description: {type: [string, 'null']}
            startedAt: {type: [string, 'null'], format: date-time}
            completedAt: {type: [string, 'null'], format: date-time}
            version: {type: integer, format: int64, minimum: 0}
...
    ProjectListApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data:
          type: array
          items: {$ref: '#/components/schemas/ProjectSummary'}
        meta: {$ref: '#/components/schemas/Meta'}
    ProjectDetailApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/ProjectDetail'}
        meta: {$ref: '#/components/schemas/Meta'}
...